// src/app/models/PlayerLeaderboard.ts
import { Member } from './Member';

export class PlayerLeaderboard {
    member?: Member;
    gamesPlayed?: number;
    wins?: number;
    losses?: number;
    draws?: number;
    goals?: number;
    primaryAssists?: number;
    secondaryAssists?: number;
    ownGoals?: number;
    plusMinus?: number;
    penaltyMinutes?: number;
    penaltyMinutesDrawn?: number;
    timeOnIce?: number;
    shifts?: number;
    evenStrengthGoals?: number;
    evenStrengthPrimaryAssists?: number;
    evenStrengthSecondaryAssists?: number;
    evenStrengthPoints?: number;
    powerPlayGoals?: number;
    powerPlayPoints?: number;
    powerPlayPrimaryAssists?: number;
    powerPlaySecondaryAssists?: number;
    shortHandedGoals?: number;
    shortHandedPrimaryAssists?: number;
    shortHandedSecondaryAssists?: number;
    shortHandedPoints?: number;
    overtimeGoals?: number;
    overtimePrimaryAssists?: number;
    overtimeSecondaryAssists?: number;
    overtimePoints?: number;
    gameWinningGoals?: number;
    gameWinningPrimaryAssists?: number;
    gameWinningSecondaryAssists?: number;
    gameWinningPoints?: number;
    emptyNetGoals?: number;
    emptyNetPrimaryAssists?: number;
    emptyNetSecondaryAssists?: number;
    emptyNetPoints?: number;
    shotsOnGoal?: number;
    shotsThatMissedNet?: number;
    shotsThatWereBlocked?: number;
    blockedShots?: number;
    offensiveZoneTakeaways?: number;
    neutralZoneTakeaways?: number;
    defensiveZoneTakeaways?: number;
    offensiveZoneGiveaways?: number;
    neutralZoneGiveaways?: number;
    defensiveZoneGiveaways?: number;
    hitsGiven?: number;
    hitsTaken?: number;
    faceoffsTaken?: number;
    faceoffsWon?: number;
    puckPossessionTime?: number;
    puckTouches?: number;
    fights?: number;
    fightsWon?: number;
}